backend/requirements.txt
alembic==1.15.2
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.12
decorator==5.2.1
docopt==0.6.2
dotenv==0.9.9
executing==2.2.0
Flask==3.1.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pydantic==2.10.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.38
stack-data==0.6.3
threadpoolctl==3.6.0
timedelta==2020.12.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.0
yarg==0.1.10
zope.interface==7.2
